requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
